requirements.txt
instabot==0.49.0
python-dotenv==0.10.3
requests==2.22.0